1차 강의 완성본

+++ src/director/director.service.ts
@@ -1,59 +1,76 @@
-import {Injectable, NotFoundException} from "@nestjs/common";
-import {CreateDirectorDto} from "./dto/create-director.dto";
-import {UpdateDirectorDto} from "./dto/update-director.dto";
-import {InjectRepository} from "@nestjs/typeorm";
-import {Director} from "./entities/director.entity";
-import {Repository} from "typeorm";
+import {Injectable, NotFoundException} from '@nestjs/common';
+import {CreateDirectorDto} from './dto/create-director.dto';
+import {UpdateDirectorDto} from './dto/update-director.dto';
+import {InjectRepository} from '@nestjs/typeorm';
+import {Director} from './entity/director.entity';
+import {Repository} from 'typeorm';
 
 @Injectable()
 export class DirectorService {
-
     constructor(
         @InjectRepository(Director)
-        private readonly directorRepository: Repository<Director>
+        private readonly directorRepository: Repository<Director>,
     ) {
     }
 
-    async create(createDirectorDto: CreateDirectorDto) {
-        return await this.directorRepository.save(createDirectorDto);
+    create(createDirectorDto: CreateDirectorDto) {
+        return this.directorRepository.save(createDirectorDto);
     }
 
-    async findAll() {
+    findAll() {
         return this.directorRepository.find();
     }
 
-    async findOne(id: number) {
-        return await this.directorRepository.findOne({
+    findOne(id: number) {
+        return this.directorRepository.findOne({
             where: {
-                id
+                id,
             },
-            relations: ["movies"]
-        });
-    }
-
-    async findOneByName(name: string) {
-        return await this.directorRepository.findOne({
-            where: {
-                name
-            }
-        });
+        })
     }
 
     async update(id: number, updateDirectorDto: UpdateDirectorDto) {
-        const director = await this.findOne(id);
+        const director = await this.directorRepository.findOne({
+            where: {
+                id,
+            },
+        });
+
         if (!director) {
-            throw new NotFoundException(`Director with id ${id} not found`);
+            throw new NotFoundException('존재하지 않는 ID의 영화입니다!');
         }
 
-        const directorUpdateFields = {
-            ...updateDirectorDto
-        };
+        await this.directorRepository.update(
+            {
+                id,
+            },
+            {
+                ...updateDirectorDto,
+            }
+        );
 
-        await this.directorRepository.update(id, directorUpdateFields);
-        return await this.findOne(id);
+        const newDirector = await this.directorRepository.findOne({
+            where: {
+                id,
+            },
+        });
+
+        return newDirector;
     }
 
-    remove(id: number) {
-        return this.directorRepository.delete(id);
+    async remove(id: number) {
+        const director = await this.directorRepository.findOne({
+            where: {
+                id,
+            },
+        });
+
+        if (!director) {
+            throw new NotFoundException('존재하지 않는 ID의 영화입니다!');
+        }
+
+        await this.directorRepository.delete(id);
+
+        return id;
     }
 }
