+++ src/user/user.service.ts
@@ -1,49 +1,110 @@
-import {Injectable, NotFoundException} from '@nestjs/common';
+import {BadRequestException, Injectable, NotFoundException} from '@nestjs/common';
 import {CreateUserDto} from './dto/create-user.dto';
 import {UpdateUserDto} from './dto/update-user.dto';
-import {InjectRepository} from "@nestjs/typeorm";
-import {Repository} from "typeorm";
-import {User} from "./entities/user.entity";
+import {Repository} from 'typeorm';
+import {InjectRepository} from '@nestjs/typeorm';
+import {User} from './entities/user.entity';
+import * as bcrypt from 'bcrypt';
+import {envVariableKeys} from 'src/common/const/env.const';
+import {ConfigService} from '@nestjs/config';
 
 @Injectable()
 export class UserService {
-
     constructor(
         @InjectRepository(User)
-        private readonly userRepository: Repository<User>
+        private readonly userRepository: Repository<User>,
+        private readonly configService: ConfigService,
     ) {
     }
 
     async create(createUserDto: CreateUserDto) {
-        return await this.userRepository.save(createUserDto);
+        const {email, password} = createUserDto;
+
+        const user = await this.userRepository.findOne({
+            where: {
+                email,
+            },
+        });
+
+        if (user) {
+            throw new BadRequestException('이미 가입한 이메일 입니다!');
+        }
+
+        const hash = await bcrypt.hash(password, this.configService.get<number>(envVariableKeys.hashRounds))
+
+        await this.userRepository.save({
+            email,
+            password: hash,
+        });
+
+        return this.userRepository.findOne({
+            where: {
+                email,
+            },
+        });
     }
 
-    async findAll() {
+    findAll() {
         return this.userRepository.find();
     }
 
     async findOne(id: number) {
         const user = await this.userRepository.findOne({
             where: {
-                id
+                id,
             }
         });
 
         if (!user) {
-            throw new NotFoundException(`User with ID ${id} not found`);
+            throw new NotFoundException('존재하지 않는 사용자입니다!');
         }
 
         return user;
     }
 
     async update(id: number, updateUserDto: UpdateUserDto) {
-        const user = await this.findOne(id);
-        await this.userRepository.update(id, updateUserDto);
-        return user;
+        const {password} = updateUserDto;
+
+        const user = await this.userRepository.findOne({
+            where: {
+                id,
+            }
+        });
+
+        if (!user) {
+            throw new NotFoundException('존재하지 않는 사용자입니다!');
+        }
+
+        const hash = await bcrypt.hash(password, this.configService.get<number>(envVariableKeys.hashRounds));
+
+        await this.userRepository.update(
+            {id},
+            {
+                ...updateUserDto,
+                password: hash,
+            },
+        );
+
+        return this.userRepository.findOne({
+            where: {
+                id,
+            },
+        });
     }
 
     async remove(id: number) {
-        await this.findOne(id);
-        return this.userRepository.delete(id);
+        const user = await this.userRepository.findOne({
+            where: {
+                id,
+            }
+        });
+
+        if (!user) {
+            throw new NotFoundException('존재하지 않는 사용자입니다!');
+        }
+
+        await this.userRepository.delete(id);
+
+        return id;
     }
 }
